Updating pow to long int

+++ 0x13-bit_manipulation/1-print_binary.c
@@ -7,7 +7,7 @@
 
 void print_binary(unsigned long int n)
 {
-	unsigned int pow = 1;
+	unsigned long int pow = 1;
 	int count = 0;
 
 	if (n == 0)
